// src/app/model/administration.ts
  
  export class Administration {
    constructor() {}
  
    persal_number?: string;
    user_id?: number;
    username?: string;
    first_name?: string;
    last_name?: string;
    email_address?: string;
    telephone_number?: string;
    role_id?: number;
    role_name?: string;


    full_names?: string;
    surname?: string;
    region_code?: number;
    organisation_code?: string;
    component_code?: number;
    sub_component_code?: number;
    paypoint_code?: number;
    responsibility_code?: number;
    objective_code?: number;
    primary_name?: string;
    primary_email?: string;
    tel_no?: string;

    post_salary_level?: number;
    system_approver?: number;
    system_recommender?: number;

}